extract and test config
