+++ internal/telemetry/metric_exporter_config_test.go
@@ -1,0 +1,68 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package telemetry
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"go.opentelemetry.io/otel/sdk/metric"
+	"go.opentelemetry.io/otel/sdk/metric/aggregation"
+	"go.opentelemetry.io/otel/sdk/metric/metricdata"
+)
+
+func TestNewExporterConfig(t *testing.T) {
+	aggregationSelector := func(metric.InstrumentKind) aggregation.Aggregation { return nil }
+	temporalitySelector := func(metric.InstrumentKind) metricdata.Temporality { return metricdata.CumulativeTemporality }
+
+	testCases := []struct {
+		name       string
+		options    []ConfigOption
+		wantConfig Config
+	}{
+		{
+			name:       "Default",
+			options:    nil,
+			wantConfig: Config{},
+		},
+		{
+			name: "WithAggregationSelector",
+			options: []ConfigOption{
+				WithAggregationSelector(aggregationSelector),
+			},
+			wantConfig: Config{},
+		},
+		{
+			name: "WithTemporalitySelector",
+			options: []ConfigOption{
+				WithTemporalitySelector(temporalitySelector),
+			},
+			wantConfig: Config{},
+		},
+	}
+	for _, tt := range testCases {
+		t.Run(tt.name, func(t *testing.T) {
+			cfg := newConfig(tt.options...)
+			// tested by TestConfigManualReaderOptions
+			cfg.AggregationSelector = nil
+			cfg.TemporalitySelector = nil
+
+			assert.Equal(t, tt.wantConfig, cfg)
+		})
+	}
+}
